+++ src/main/java/com/bmilab/backend/global/utils/ExcelGenerator.java
@@ -2,6 +2,7 @@ package com.bmilab.backend.global.utils;
 
 import lombok.val;
 import org.apache.poi.ss.usermodel.CellStyle;
+import org.apache.poi.ss.usermodel.FillPatternType;
 import org.apache.poi.ss.usermodel.Font;
 import org.apache.poi.ss.usermodel.IndexedColors;
 import org.apache.poi.xssf.streaming.SXSSFCell;
@@ -63,7 +64,8 @@ public class ExcelGenerator {
 
         CellStyle headerCellStyle = workbook.createCellStyle();
         headerCellStyle.setFont(headerFont);
-        headerCellStyle.setFillBackgroundColor(IndexedColors.DARK_BLUE.getIndex());
+        headerCellStyle.setFillForegroundColor(IndexedColors.DARK_BLUE.getIndex());
+        headerCellStyle.setFillPattern(FillPatternType.SOLID_FOREGROUND);
 
         SXSSFRow headerRow = sheet.createRow(HEADER_ROW);
